A snippet of code from my in browser RPG game
I'm making an in-browser card-based RPG game where you battle monsters. This code snippet is what I'm using to handle the monsters you battle in the game. I'm using a React component that pulls stats and other information. To keep the monsters scalable I made the component pull from a list of classes that and be edited at any time. The component also has a few events linked to buttons that simulate damage dealt and applied status effects.